requirements.txt
aioamqp==0.13.0
asgiref==3.2.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-rabbitmq==1.1.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.5
django-channels==0.7.0
django-cors-headers==3.1.0
django-rest==0.0.1
djangorestframework==3.10.2
essential-generators==0.9.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jsonschema==3.0.2
lazy-object-proxy==1.4.2
msgpack==0.6.1
oauthlib==3.1.0
openapi-core==0.11.0
openapi-spec-validator==0.2.8
pamqp==2.3.0
Pillow==6.1.0
pycparser==2.19
PyHamcrest==1.9.0
pyrsistent==0.15.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
strict-rfc3339==0.7
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
zope.interface==4.6.0
